contribuitions/course-automation/ntas-amar/requirements.txt
autosemver==0.5.5
certifi==2020.12.5
chardet==4.0.0
Deprecated==1.2.12
dulwich==0.19.16
idna==2.10
PyGithub==1.54.1
PyJWT==1.7.1
PyPDF2==1.26.0
python-magic==0.4.22
refextract
requests==2.25.1
six==1.15.0
Unidecode==1.2.0
urllib3==1.26.4
wrapt==1.12.1
